+++ resources/views/index.blade.php
@@ -5,18 +5,6 @@ $config = [
     'locales' => config('app.locales'),
     'githubAuth' => config('services.github.client_id'),
 ];
-
-$polyfills = [
-    'Promise',
-    'Object.assign',
-    'Object.values',
-    'Array.prototype.find',
-    'Array.prototype.findIndex',
-    'Array.prototype.includes',
-    'String.prototype.includes',
-    'String.prototype.startsWith',
-    'String.prototype.endsWith',
-];
 @endphp
 <!DOCTYPE html>
 <html lang="{{ app()->getLocale() }}">
@@ -35,9 +23,6 @@ $polyfills = [
   <script>
     window.config = @json($config);
   </script>
-
-  {{-- Polyfill JS features via polyfill.io --}}
-  <script src="https://cdn.polyfill.io/v2/polyfill.min.js?features={{ implode(',', $polyfills) }}"></script>
 
   {{-- Load the application scripts --}}
   <script src="{{ mix('dist/js/app.js') }}"></script>