+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Update npm dependencies
 - Removed yarn.lock in favor of package-lock.json
 - Fixed `laravel-mix` (HMR, versioning, build cleanup), thanks to [TemaSM](https://github.com/TemaSM)
+- Removed JS polyfills, if you want to support IE11 see [laravel-mix#436](https://github.com/JeffreyWay/laravel-mix/issues/436) 
 
 ## 4.3.0 - 2018-10-21